feat: usage redis timer for lobby
